Fixed Bug Causing Multiple Cards to be displayed on Step /Page 3
I utilised the jQuery on and off event handlers in order to restrict click events triggers on the cards to once per page. This resulted in only one call to the function stepThree being possible per page/step. Using the on binding also allowed me to reduce the number of global variables that I had to declare