All major Js functionality has been added to page .
Added functionality for the h1 element to display the number of  drops to use dependent on beard length

Bug fixed which caused multiple duplicated elements to be appended to the Step 3 Page. This occured when multiple elements were clicked on Step 1 and Step 2 before they fully faded out. I utilised jQuery on / off methods to restrict events to one execution per click of an element class to resolve this.